==> 00_empirical-n-cont/1_calc-aonr.R <==
# created 2/12/21
# gina
# purpose: mike thinks I can get some kind of info from aonr stuff
# updated: 3/19/2021 separated aonr calcs from npct calcs
#          3/24/2021 - do calcs on sim data also
#         8/3/2021 - moved sim calcs to own folder, cleaned up

rm(list = ls())

library(tidysawyer2)
library(tidyverse)
library(nlraa)

# functions ---------------------------------------------------------------------

tst <- 
  ilia_yields %>%
  filter(site == "ames",
         year %in% c(2000),
         rotation == "cc") 


tst.m <- nls(yield_kgha ~ (a + b * nrate_kgha + c * I(nrate_kgha^2)) *
                                          (nrate_kgha <= -0.5 * b/c) +
                                          (a + I(-b^2/(4 * c))) * (nrate_kgha > -0.5 * b/c),
                                        start = list(a = 1.37,
                                                     b = 0.0215,
                                                     c = -0.0000568),
                                        control = list(maxiter = 1000),
                                        data = tst)

coef(tst.m)


  
#--map it, will those starting values work for everything?

qpfit_fun <- function(tst = data){
  
  tst.m <- nls(yield_kgha ~ (a + b * nrate_kgha + c * I(nrate_kgha^2)) *
                 (nrate_kgha <= -0.5 * b/c) +
                 (a + I(-b^2/(4 * c))) * (nrate_kgha > -0.5 * b/c),
               start = list(a = 1.37,
                            b = 0.0215,
                            c = -0.0000568),
               control = list(maxiter = 1000),
               data = tst)
  return(tst.m)
  
}

#--specifically to get coefficients and aonr
qpcoefs_fun <- function(tst = data){
  
  tst.m <- nls(yield_kgha ~ (a + b * nrate_kgha + c * I(nrate_kgha^2)) *
                 (nrate_kgha <= -0.5 * b/c) +
                 (a + I(-b^2/(4 * c))) * (nrate_kgha > -0.5 * b/c),
               start = list(a = 1.37,
                            b = 0.0215,
                            c = -0.0000568),
               control = list(maxiter = 1000),
               data = tst)
  
tst.coef <- 
    coef(tst.m) %>% 
    as.data.frame() %>% 
    rownames_to_column() %>% 
    as_tibble() %>% 
    rename("coef" = 2) %>% 
    pivot_wider(names_from = rowname, values_from = coef)
  
  return(tst.coef)
  
}

qpfit_fun(tst = tst)
qpcoefs_fun(tst = tst)


# do on all sites, observed data --------------------------------------------
#--note: what happens in years where no plateau is reached?

dat <- 
  ilia_yields

sy <- 
  dat %>% 
  select(site, year, rotation) %>% 
  distinct()

# coefficients ------------------------------------------------------------

d.coefs <- 
  dat %>% 
  group_by(site, year, rotation) %>%
  nest() %>%
  mutate(model = map(data, possibly(qpcoefs_fun, NULL))) %>% 
  rowwise() %>% 
  filter(!is.null(model)) %>%
  unnest(cols = c(model)) 


# aonrs -------------------------------------------------------------------

d.aonrs <- 
  dat %>% 
  group_by(site, year, rotation) %>%
  nest() %>%
  mutate(model = map(data, possibly(qpcoefs_fun, NULL))) %>% 
  rowwise() %>% 
  filter(!is.null(model)) %>%
  unnest(cols = c(model)) %>% 
  mutate(aonr_kgha = -0.5 * (b/c)) %>% 
#  select(site, year, rotation, aonr_kgha) %>% 
  mutate(rotation = paste0("aonr_", rotation),
         aonr_kgha = round(aonr_kgha, 0)) %>% 
  rename("aonr_rot" = rotation)  

d.aonrs %>% 
  write_csv("00_empirical-n-cont/dat_aonrs.csv")



# combine for manu table --------------------------------------------------

d.manu <- 
  dat %>% 
  group_by(site, year, rotation) %>%
  nest() %>%
  mutate(model = map(data, possibly(qpcoefs_fun, NULL))) %>% 
  rowwise() %>% 
  filter(!is.null(model)) %>%
  unnest(cols = c(model)) %>% 
  mutate(aonr_kgha = -0.5 * (b/c)) %>% 
  mutate(aonr_kgha = round(aonr_kgha, 0))

d.manu %>% 
  ungroup() %>% 
  right_join(sy) %>% 
  left_join(ilia_siteinfo %>% select(site, manu_id)) %>% 
  select(manu_id, year, rotation, a, b, c, aonr_kgha) %>% 
  arrange(manu_id, year, rotation) %>% 
  write_csv("00_empirical-n-cont/dat_manu-params-aonrs.csv")
